Reworked to re-export the original component
