--- packages/patternfly-3/patternfly-react-wooden-tree/src/components/TreeView/TreeView.fixtures.js
@@ -1,33 +1,0 @@
-export const data = [
-  {
-    text: 'Parent 1',
-    icon: 'fa fa-folder',
-    nodes: [
-      {
-        text: 'Child 1',
-        nodes: [{ text: 'Grandchild 1' }, { text: 'Grandchild 2' }]
-      },
-      {
-        text: 'Child 2',
-        nodes: [{ text: 'Grandchild 1' }, { text: 'Grandchild 2' }]
-      }
-    ]
-  },
-  {
-    text: 'Parent 2',
-    icon: 'fa fa-folder',
-    nodes: [{ text: 'Child 1' }]
-  },
-  {
-    text: 'Parent 3',
-    icon: 'fa fa-folder'
-  },
-  {
-    text: 'Parent 4',
-    icon: 'fa fa-folder'
-  },
-  {
-    text: 'Parent 5',
-    icon: 'fa fa-folder'
-  }
-];
