+++ packages/patternfly-3/patternfly-react-wooden-tree/src/index.js
@@ -1,1 +1,1 @@
-export * from './components/TreeView';
+export { default as TreeView } from './components/TreeView';
